backend/requirements.txt
transformers==4.25.1
datasets
sentencepiece
torch
pytorch-lightning==2.0.7
wandb
sconf
donut-python
timm==0.5.4